requirements.txt
django
django-rest-framework
django-cors-headers
django-simple-history
preview-generator
tabulate
coverage